Initial Push away from PlayerPrefs
This push includes my work to start removing references to PlayerPrefs due to the fact that they weren't meant to hold this much of this kind of data. It will now save to a file the default unity application path.

There will be 2 binary files made:
traits.gd
characterStats.gd

This build is not tested very much and will probably break under stress. Initial tests show it saving and loading properly.